+++ charts/kargo/templates/kubernetes-webhooks-server/cluster-role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: kargo-webhooks-server
   labels:
     {{- include "kargo.labels" . | nindent 4 }}
-    {{- include "kargo.webhooksServer.labels" . | nindent 4 }}
+    {{- include "kargo.kubernetesWebhooksServer.labels" . | nindent 4 }}
 rules:
 - apiGroups:
   - ""
@@ -78,7 +78,7 @@ metadata:
   name: kargo-webhooks-server-ns-controller
   labels:
     {{- include "kargo.labels" . | nindent 4 }}
-    {{- include "kargo.webhooksServer.labels" . | nindent 4 }}
+    {{- include "kargo.kubernetesWebhooksServer.labels" . | nindent 4 }}
 rules:
 - apiGroups:
   - kargo.akuity.io
@@ -99,7 +99,7 @@ metadata:
   name: kargo-webhooks-server-generic-gc
   labels:
     {{- include "kargo.labels" . | nindent 4 }}
-    {{- include "kargo.webhooksServer.labels" . | nindent 4 }}
+    {{- include "kargo.kubernetesWebhooksServer.labels" . | nindent 4 }}
 rules:
 - apiGroups:
   - kargo.akuity.io
